Add files via upload
Glitchy Bird Reloaded is my artistic representation of stress and failure. The interactive p5js sketch will bombard the viewer nonstop with barriers that the viewer must avoid using the mouse to move their “pulse” out of harm’s way. 

As if a beating heart; the viewer’s mouse guided avatar beats in and out in order to literally represent our varying ability day to day in avoiding life's obstacles. 

The moment the viewer is unable to avoid an obstacle; time will for a moment freeze as you are confronted with a pixelated version of you failing pulled from the webcam.
	Through the Glitch Art aesthetic, I am expressing the utter frustration and stress felt from the feeling of being overwhelmed in life. It also confronts the main source of one’s stress and frustrations, one’s own failures. 
In being confronted with your own image every time you stumble upon a barrier in “Glitchy Bird Reloaded” you are immediately confronted with the driving force behind your failures.